requirements.txt
pandas
requests
apache-airflow
hdfs
